model/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
numpy
opencv-python-headless
Pillow
PyMuPDF
pymongo
python-dotenv
torch==2.1.0+cpu
torchvision==0.16.0+cpu